rent en el backend

+++ backend/rent/views.py
@@ -6,39 +6,67 @@ from core.permissions import IsAdmin
 from .models import Rent
 
 class RentView(viewsets.GenericViewSet):
-    # permission_classes = (IsAuthenticated,)
-    print('hola1')
-    # def rent(self, request, slot_id):
+    def rent(self, request, slot_id):
+        bearer = request.headers['Authorization'].split()
         
-    #     username = request.user
-    #     serializer_context = { 'username': username, 'slot_id': slot_id }
-    #     serializer = RentSerializer.rent(context=serializer_context)
-    #     return Response(RentSerializer.to_rent(serializer))
+        serializer_context_user = {
+            'token': bearer[1]
+        }
+        serializer = RentSerializer.usertoken(
+        context=serializer_context_user)
+        username = serializer[0]
+        serializer_context = { 'username': username, 'slot_id': slot_id }
+        serializer = RentSerializer.rent(context=serializer_context)
+        return Response(RentSerializer.to_rent(serializer))
 
     def getOneRent(self, request):
-        # print('hola2')
-        username = request.user
+        bearer = request.headers['Authorization'].split()
+        serializer_context_user = {
+            'token': bearer[1]
+        }
+        serializer_user = RentSerializer.usertoken(
+        context=serializer_context_user)
+        username = serializer_user[0]
         serializer_context = { 'username': username }
         serializer = RentSerializer.getOneRent(context=serializer_context)
         return Response(RentSerializer.to_rent(serializer))
 
     def bringbackBicis(self, request):
+        bearer = request.headers['Authorization'].split()
+        serializer_context_user = {
+            'token': bearer[1]
+        }
+        serializer_user = RentSerializer.usertoken(
+        context=serializer_context_user)
+        username = serializer_user[0]
         data = request.data['bici']
-        username = request.user
         serializer_context = {'username': username, 'slot_id': data['end_slot'], 'bici_id': data['bici_id']}
         serializer = RentSerializer.bringbackBicis(context=serializer_context)
         return Response(RentSerializer.to_rent(serializer))
 
 class RentAdminView(viewsets.GenericViewSet):
-    permission_classes = [IsAdmin]
 
     def getAllRents(self, request):
-        data = Rent.objects.all()
-        serializer = RentSerializer(data, many=True)
+        bearer = request.headers['Authorization'].split()
+        serializer_context_user = {
+            'token': bearer[1]
+        }
+        serializer_user = RentSerializer.usertoken(
+        context=serializer_context_user)
+        if serializer_user[2] == 'admin':
+            data = Rent.objects.all()
+            serializer = RentSerializer(data, many=True)
         return Response(serializer.data)
 
     def deleteRent(self, request, id):
-        rent = Rent.objects.get(id=id)
-        rent.delete()
+        bearer = request.headers['Authorization'].split()
+        serializer_context_user = {
+            'token': bearer[1]
+        }
+        serializer_user = RentSerializer.usertoken(
+        context=serializer_context_user)
+        if serializer_user[2] == 'admin':
+            rent = Rent.objects.get(id=id)
+            rent.delete()
         return Response({'data': 'Rent deleted successfully'})
